fixed fix-list view error
In an attempt to try to modularize the data sent back when a fix status
is saved, I introduced a couple of bugs that I did not commit. Example:
In creating a form for each fix the select options stopped sending post
values. This has now been fixed by adding "form = fixNumber-formID". but
there is still a bug where the checkbox buttons are out of synch:
checking the button works, but unchecking it needs the form to be
submitted twice. Will fix this tomorrow