completed itemList view
items are viewed in tables under todo, notes and reminders. Each item
can be edited (owner, for, type, and duedate ) or deleted. For this I'll
be using modals, and the same functionality as the Fix Testcomplete,
because I need a way to tell the modal which item it is being called for