changed addFix functionality
Now when a user clicks the button, a javascript function handles how the
modal comes up, depending on whether the fix is being added to a sprint
or just being added. This functionality will change as we will  just
expect rady fixes to be pulled directly from the onefix server. i.e Test
Lead should not have to add them